+++ src/main/java/cart/domain/PointEarningPolicy.java
@@ -1,4 +1,4 @@
-package cart.application;
+package cart.domain;
 
 import cart.domain.Point;
